move: helpers out of common.py
