+++ benchmark/python-rs-driver/concurrent_complex_select.py
@@ -1,13 +1,12 @@
 import asyncio
 import sys
 
-from scylla.session import Session
-from scylla.statement import PreparedStatement
-
 from common import (
-    connect,
     COMPLEX_SELECT_QUERY,
 )
+from python_rs_helpers import connect
+from scylla.session import Session
+from scylla.statement import PreparedStatement
 
 
 async def select_data(
