+++ benchmark/python-rs-driver/insert.py
@@ -1,16 +1,13 @@
+import asyncio
 import sys
 
-import asyncio
-
-from scylla.session import Session
-from scylla.statement import PreparedStatement
-from scylla.statement import Statement
-
 from common import (
-    connect,
     SIMPLE_INSERT_QUERY,
     get_simple_data,
 )
+from python_rs_helpers import connect
+from scylla.session import Session
+from scylla.statement import PreparedStatement, Statement
 
 
 async def test(session: Session, prepared: PreparedStatement, cnt: int):
